+++ internal/db/repositories.sql.go
@@ -47,7 +47,7 @@ type CreateRepositoryParams struct {
 	RepoID        int64          `json:"repo_id"`
 	IsPrivate     bool           `json:"is_private"`
 	IsFork        bool           `json:"is_fork"`
-	WebhookID     sql.NullInt32  `json:"webhook_id"`
+	WebhookID     sql.NullInt64  `json:"webhook_id"`
 	WebhookUrl    string         `json:"webhook_url"`
 	DeployUrl     string         `json:"deploy_url"`
 	CloneUrl      string         `json:"clone_url"`
@@ -458,7 +458,7 @@ type UpdateRepositoryParams struct {
 	RepoID        int64          `json:"repo_id"`
 	IsPrivate     bool           `json:"is_private"`
 	IsFork        bool           `json:"is_fork"`
-	WebhookID     sql.NullInt32  `json:"webhook_id"`
+	WebhookID     sql.NullInt64  `json:"webhook_id"`
 	WebhookUrl    string         `json:"webhook_url"`
 	DeployUrl     string         `json:"deploy_url"`
 	Provider      string         `json:"provider"`
@@ -528,7 +528,7 @@ type UpdateRepositoryByIDParams struct {
 	RepoName      string         `json:"repo_name"`
 	IsPrivate     bool           `json:"is_private"`
 	IsFork        bool           `json:"is_fork"`
-	WebhookID     sql.NullInt32  `json:"webhook_id"`
+	WebhookID     sql.NullInt64  `json:"webhook_id"`
 	WebhookUrl    string         `json:"webhook_url"`
 	DeployUrl     string         `json:"deploy_url"`
 	Provider      string         `json:"provider"`
